Unify datum parsing
